+++ NEWS.md
@@ -6,7 +6,7 @@
 * Change pandoc output to use inline images if there is no caption
 * Use Documenter.jl for documentation
 * Add chunk option `hold`, replaces results = "hold". This way you can use e.g. `hold = true, results=raw`.
-
+* Methods for setting and restoring default chunk options for documents.
 
 ### v0.1.0