+++ NEWS.md
@@ -7,6 +7,8 @@
 * Use Documenter.jl for documentation
 * Add chunk option `hold`, replaces results = "hold". This way you can use e.g. `hold = true, results=raw`.
 * Methods for setting and restoring default chunk options for documents.
+* New output options `md2pdf` and `md2html`, both use pandoc to output pdf and
+  html files directly with own templates.
 
 ### v0.1.0